# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/admin/users/{id}/change-email": {
      "post": {
        "description": "Cambia el email del usuario y genera una nueva contraseña temporal. Revoca todas las sesiones activas y envía un welcome email con las nuevas credenciales al nuevo email. CRÍTICO: Solo para administradores. El usuario debe hacer login con la nueva contraseña temporal.",
        "operationId": "UserIdentityController_changeEmail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email cambiado exitosamente. Se envió welcome email con credenciales",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email cambiado exitosamente. Se envió un email de bienvenida con las nuevas credenciales a nuevo@email.com."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "El nuevo email ya está en uso"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "summary": "Cambiar email de usuario (SOLO ADMIN)",
        "tags": [
          "auth-administration"
        ]
      }
    },
...
    "/api/audits/{id}": {
      "get": {
        "description": "Retorna los datos completos de un audit específico mediante su ID único.",
        "operationId": "AuditsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recurso encontrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado - Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos - No tiene acceso a este recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener auditoría por ID",
        "tags": [
          "Audits - Auditorías"
        ]
      },
      "patch": {
        "description": "Solo se pueden actualizar auditorías en estado DRAFT",
        "operationId": "AuditsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuditDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recurso actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado - Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos - No tiene acceso a este recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Actualizar auditoría",
        "tags": [
          "Audits - Auditorías"
        ]
      },
      "delete": {
        "description": "Solo se pueden eliminar auditorías en estado DRAFT",
        "operationId": "AuditsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado - Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos - No tiene acceso a este recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Eliminar auditoría (soft delete)",
        "tags": [
          "Audits - Auditorías"
        ]
      }
    },
...
    "/api/audits/{auditId}/assignments": {
      "post": {
        "description": "Asigna un usuario al equipo de auditoría con un rol específico (LEAD_AUDITOR, AUDITOR, AUDITEE, OBSERVER)",
        "operationId": "AuditAssignmentsController_assign",
        "parameters": [
          {
            "name": "auditId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recurso creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditAssignmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado - Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos - No tiene acceso a este recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "El recurso ya existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Asignar miembro a auditoría",
        "tags": [
          "Audits - Asignaciones de Miembros"
        ]
      },
      "get": {
        "description": "Obtiene todos los miembros activos asignados a la auditoría",
        "operationId": "AuditAssignmentsController_findAll",
        "parameters": [
          {
            "name": "auditId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditAssignmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado - Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos - No tiene acceso a este recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Listar miembros de auditoría",
        "tags": [
          "Audits - Asignaciones de Miembros"
        ]
      }
    },
    "/api/audits/{auditId}/assignments/{assignmentId}": {
      "delete": {
        "description": "Desactiva la asignación de un miembro (soft delete)",
        "operationId": "AuditAssignmentsController_remove",
        "parameters": [
          {
            "name": "auditId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditAssignmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado - Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos - No tiene acceso a este recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Remover miembro de auditoría",
        "tags": [
          "Audits - Asignaciones de Miembros"
        ]
      }
    },
...
  },
  "info": {
    "title": "ATR API",
    "description": "API de auditorías, plantillas, frameworks de madurez y gestión de usuarios",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChangeUserEmailDto": {
        "type": "object",
        "properties": {
          "newEmail": {
            "type": "string",
            "description": "Nuevo email del usuario",
            "example": "nuevo.email@example.com",
            "maxLength": 100
          }
        },
        "required": [
          "newEmail"
        ]
      },
...
      "AuditResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "code": {
            "type": "string",
            "example": "AUD-2024-001"
          },
          "name": {
            "type": "string",
            "example": "Auditoría ISO 27001 - Empresa XYZ Q1 2024"
          },
          "description": {
            "type": "object",
            "example": "Evaluación anual de seguridad de información",
            "nullable": true
          },
          "templateId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "organizationId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "frameworkId": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440003",
            "nullable": true
          },
          "parentAuditId": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440004",
            "nullable": true
          },
          "revisionNumber": {
            "type": "number",
            "example": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_PROGRESS",
              "CLOSED",
              "ARCHIVED"
            ],
            "example": "DRAFT"
          },
          "startDate": {
            "type": "object",
            "example": "2024-02-01T00:00:00.000Z",
            "nullable": true
          },
          "endDate": {
            "type": "object",
            "example": "2024-02-28T23:59:59.999Z",
            "nullable": true
          },
          "actualStartDate": {
            "type": "object",
            "example": "2024-02-01T09:00:00.000Z",
            "nullable": true
          },
          "closedAt": {
            "type": "object",
            "example": "2024-02-28T18:00:00.000Z",
            "nullable": true
          },
          "overallScore": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "maturityLevel": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "createdBy": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440005",
            "nullable": true
          },
          "updatedBy": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440005",
            "nullable": true
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "templateId",
          "organizationId",
          "frameworkId",
          "parentAuditId",
          "revisionNumber",
          "status",
          "startDate",
          "endDate",
          "actualStartDate",
          "closedAt",
          "overallScore",
          "maturityLevel",
          "createdAt",
          "updatedAt",
          "createdBy",
          "updatedBy"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica si la operación fue exitosa (siempre false para errores)",
            "example": false,
            "default": false
          },
          "statusCode": {
            "type": "number",
            "description": "Código de estado HTTP del error",
            "example": 400,
            "enum": [
              400,
              401,
              403,
              404,
              409,
              422,
              500
            ]
          },
          "message": {
            "description": "Mensaje de error. Puede ser un string simple o array de strings para errores de validación",
            "oneOf": [
              {
                "type": "string",
                "example": "Usuario no encontrado"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "email debe ser un email válido",
                  "password es requerido"
                ]
              }
            ]
          },
          "error": {
            "type": "string",
            "description": "Nombre del error HTTP",
            "example": "Bad Request",
            "enum": [
              "Bad Request",
              "Unauthorized",
              "Forbidden",
              "Not Found",
              "Conflict",
              "Unprocessable Entity",
              "Internal Server Error"
            ]
          },
          "errors": {
            "type": "array",
            "description": "Array de errores adicionales (usado en errores de importación masiva, etc.)",
            "items": {
              "type": "object"
            }
          },
          "validationErrors": {
            "type": "object",
            "description": "Errores de validación agrupados por campo (generado por class-validator)",
            "example": {
              "email": [
                "debe ser un email válido"
              ],
              "password": [
                "es requerido",
                "debe tener al menos 8 caracteres"
              ]
            }
          },
          "summary": {
            "type": "object",
            "description": "Resumen de errores (usado en operaciones batch)"
          },
          "totalErrors": {
            "type": "number",
            "description": "Número total de errores encontrados",
            "example": 5
          },
          "details": {
            "type": "object",
            "description": "Detalles adicionales del error (solo en desarrollo)"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp ISO 8601 de cuando ocurrió el error",
            "example": "2024-01-30T10:30:00.000Z"
          },
          "path": {
            "type": "string",
            "description": "Ruta del endpoint que generó el error",
            "example": "/api/users/123"
          },
          "method": {
            "type": "string",
            "description": "Método HTTP que generó el error",
            "example": "POST",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE"
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "error",
          "timestamp"
        ]
      },
...
      "UpdateAuditDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre descriptivo de la auditoría",
            "example": "Auditoría ISO 27001 - Empresa XYZ Q1 2024 (Actualizado)",
            "minLength": 3,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Descripción detallada de la auditoría",
            "example": "Evaluación actualizada con nuevo alcance"
          },
          "frameworkId": {
            "type": "string",
            "description": "ID del framework de madurez a utilizar",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "startDate": {
            "type": "string",
            "description": "Fecha de inicio planificada",
            "example": "2024-02-01"
          },
          "endDate": {
            "type": "string",
            "description": "Fecha de fin planificada",
            "example": "2024-02-28"
          }
        }
      },
...
      "AuditAssignmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "auditId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "userId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "role": {
            "type": "string",
            "enum": [
              "LEAD_AUDITOR",
              "AUDITOR",
              "AUDITEE",
              "OBSERVER"
            ],
            "example": "LEAD_AUDITOR"
          },
          "assignedStandardIds": {
            "example": [
              "550e8400-e29b-41d4-a716-446655440003",
              "550e8400-e29b-41d4-a716-446655440004"
            ],
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "object",
            "example": "Responsable de sección A.5",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "createdBy": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440005",
            "nullable": true
          },
          "updatedBy": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440005",
            "nullable": true
          }
        },
        "required": [
          "id",
          "auditId",
          "userId",
          "role",
          "assignedStandardIds",
          "notes",
          "isActive",
          "createdAt",
          "updatedAt",
          "createdBy",
          "updatedBy"
        ]
      },
      "AssignMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID del usuario a asignar",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "role": {
            "type": "string",
            "description": "Rol del miembro en la auditoría",
            "enum": [
              "LEAD_AUDITOR",
              "AUDITOR",
              "AUDITEE",
              "OBSERVER"
            ],
            "example": "AUDITOR"
          },
          "assignedStandardIds": {
            "description": "IDs de estándares específicos asignados a este miembro (opcional). Si es null, tiene acceso a todos.",
            "example": [
              "550e8400-e29b-41d4-a716-446655440001",
              "550e8400-e29b-41d4-a716-446655440002"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "description": "Notas sobre la asignación (ej: \"Responsable de sección A.5\")",
            "example": "Experto en controles de acceso"
          }
        },
        "required": [
          "userId",
          "role"
        ]
      },
...
    }
  }
}